cleanup and refactoring of bg api tests
